Revert "Commit 16 (#8)"
This reverts commit 6823a227b74b6a7f4bb786336808df518ad37512.
