+++ app/build.gradle
@@ -30,15 +30,9 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
-
-    buildFeatures {
-        viewBinding true
-    }
 }
 
 dependencies {
-    implementation "io.insert-koin:koin-android:3.3.0"
-    implementation "androidx.core:core-ktx:1.9.0"
     implementation 'io.reactivex.rxjava3:rxjava:3.1.8'
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
