+++ app/src/main/java/com/my/playlistmaker/App.kt
@@ -1,0 +1,38 @@
+package com.my.playlistmaker
+
+import android.app.Application
+import android.content.SharedPreferences
+import androidx.appcompat.app.AppCompatDelegate
+
+const val SETTING_EXAMPLE_PREFERENCES = "setting_example_preferences"
+const val SWITCH_KEY = "key_for_switch"
+const val HISTORY_EXAMPLE_PREFERENCES = "history_example_preferences"
+const val HISTORY_KEY = "key_for_history_switch"
+
+class App : Application() {
+
+    private var darkTheme = false
+    lateinit var sharedPrefs: SharedPreferences
+
+    override fun onCreate() {
+        super.onCreate()
+        sharedPrefs = getSharedPreferences(SETTING_EXAMPLE_PREFERENCES, Application.MODE_PRIVATE)
+        darkTheme = sharedPrefs.getBoolean(SWITCH_KEY, darkTheme)
+        switchTheme(darkTheme)
+    }
+
+    fun switchTheme(darkThemeEnabled: Boolean) {
+        darkTheme = darkThemeEnabled
+        AppCompatDelegate.setDefaultNightMode(
+            if (darkThemeEnabled) {
+                AppCompatDelegate.MODE_NIGHT_YES
+            } else {
+                AppCompatDelegate.MODE_NIGHT_NO
+            }
+        )
+        sharedPrefs = getSharedPreferences(SETTING_EXAMPLE_PREFERENCES, Application.MODE_PRIVATE)
+        sharedPrefs.edit()
+            .putBoolean(SWITCH_KEY, darkTheme)
+            .apply()
+    }
+}
