+++ app/src/main/java/com/my/playlistmaker/LibraryActivity.kt
@@ -1,0 +1,11 @@
+package com.my.playlistmaker
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class LibraryActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_library)
+    }
+}
