+++ app/src/main/java/com/my/playlistmaker/data/Player.kt
@@ -1,0 +1,10 @@
+package com.my.playlistmaker.data
+
+interface Player {
+    fun playbackControl()
+    fun onCreate(trackURL:String)
+    fun onDestroy()
+    fun play()
+    fun pause()
+
+}
