--- app/src/main/java/com/my/playlistmaker/data/api/NetworkClient.kt
@@ -1,5 +1,0 @@
-package com.my.playlistmaker.data.api
-
-interface NetworkClient {
-    fun doRequest(dto: Any): Response
-}
