--- app/src/main/java/com/my/playlistmaker/data/api/Response.kt
@@ -1,5 +1,0 @@
-package com.my.playlistmaker.data.api
-
-open class Response() {
-    var resultCode = 0
-}
