--- app/src/main/java/com/my/playlistmaker/data/api/TracksRepository.kt
@@ -1,8 +1,0 @@
-package com.my.playlistmaker.data.api
-
-import com.my.playlistmaker.Track
-import com.my.playlistmaker.domain.api.Resource
-
-interface TracksRepository {
-    fun searchTracks(expression: String): Resource<List<Track>>
-}
