--- app/src/main/java/com/my/playlistmaker/data/api/TracksSearchRequest.kt
@@ -1,3 +1,0 @@
-package com.my.playlistmaker.data.api
-
-data class TracksSearchRequest(val expression: String)
