--- app/src/main/java/com/my/playlistmaker/data/search/SearchHistoryRepository.kt
@@ -1,9 +1,0 @@
-package com.my.playlistmaker.data.search
-
-import com.my.playlistmaker.Track
-
-interface SearchHistoryRepository {
-    fun getList(): List<Track>
-    fun addTrack(track: Track)
-    fun clearList()
-}
