--- app/src/main/java/com/my/playlistmaker/data/settings/SettingsRepository.kt
@@ -1,6 +1,0 @@
-package com.my.playlistmaker.data.settings
-
-interface SettingsRepository {
-    fun getThemeState(): Boolean
-    fun setThemeState(darkThemeEnabled: Boolean)
-}
