--- app/src/main/java/com/my/playlistmaker/data/sharing/SharingRepository.kt
@@ -1,7 +1,0 @@
-package com.my.playlistmaker.data.sharing
-
-interface SharingRepository {
-    fun share()
-    fun support()
-    fun terms()
-}
