--- app/src/main/java/com/my/playlistmaker/domain/api/Resource.kt
@@ -1,6 +1,0 @@
-package com.my.playlistmaker.domain.api
-
-sealed class Resource<T>(val data: T? = null, val message: String? = null) {
-    class Success<T>(data: T): Resource<T>(data)
-    class Error<T>(message: String, data: T? = null): Resource<T>(data, message)
-}
