--- app/src/main/java/com/my/playlistmaker/domain/api/TracksInteractor.kt
@@ -1,11 +1,0 @@
-package com.my.playlistmaker.domain.api
-
-import com.my.playlistmaker.Track
-
-interface TracksInteractor {
-    fun searchTracks(expression: String, consumer: TracksConsumer)
-
-    interface TracksConsumer {
-        fun consume(foundTracks: List<Track>?, errorMessage: String?)
-    }
-}
