--- app/src/main/java/com/my/playlistmaker/domain/api/impl/TracksInteractorImpl.kt
@@ -1,21 +1,0 @@
-package com.my.playlistmaker.domain.api.impl
-
-import com.my.playlistmaker.data.api.TracksRepository
-import com.my.playlistmaker.domain.api.Resource
-import com.my.playlistmaker.domain.api.TracksInteractor
-import java.util.concurrent.Executors
-
-class TracksInteractorImpl(private val repository: TracksRepository): TracksInteractor {
-
-    private val executor = Executors.newCachedThreadPool()
-
-    override fun searchTracks(expression: String, consumer: TracksInteractor.TracksConsumer) {
-        executor.execute {
-            when(val resource = repository.searchTracks(expression)) {
-                is Resource.Success -> { consumer.consume(resource.data, null) }
-                is Resource.Error -> { consumer.consume(null, resource.message) }
-            }
-        }
-    }
-
-}
