--- app/src/main/java/com/my/playlistmaker/domain/search/SearchHistoryInteractor.kt
@@ -1,10 +1,0 @@
-package com.my.playlistmaker.domain.search
-
-import com.my.playlistmaker.Track
-
-
-interface SearchHistoryInteractor {
-    fun getList(): List<Track>
-    fun addTrack(track: Track)
-    fun clearList()
-}
