--- app/src/main/java/com/my/playlistmaker/domain/settings/SettingsInteractor.kt
@@ -1,6 +1,0 @@
-package com.my.playlistmaker.domain.settings
-
-interface SettingsInteractor {
-    fun getThemeState(): Boolean
-    fun setThemeState(darkThemeEnabled: Boolean)
-}
