--- app/src/main/java/com/my/playlistmaker/domain/sharing/SharingInteractor.kt
@@ -1,7 +1,0 @@
-package com.my.playlistmaker.domain.sharing
-
-interface SharingInteractor {
-    fun share()
-    fun support()
-    fun terms()
-}
