--- app/src/main/java/com/my/playlistmaker/presentation/LibraryActivity.kt
@@ -1,16 +1,0 @@
-package com.my.playlistmaker.presentation
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.my.playlistmaker.databinding.ActivityLibraryBinding
-
-class LibraryActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityLibraryBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityLibraryBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-    }
-}
