--- app/src/main/java/com/my/playlistmaker/presentation/player/PlayerState.kt
@@ -1,8 +1,0 @@
-package com.my.playlistmaker.presentation.player
-
-enum class PlayerState {
-        STATE_DEFAULT,
-        STATE_PREPARED,
-        STATE_PLAYING,
-        STATE_PAUSED,
-}
