--- app/src/main/java/com/my/playlistmaker/presentation/search/RecyclerViewClickListener.kt
@@ -1,7 +1,0 @@
-package com.my.playlistmaker.presentation.search
-
-import com.my.playlistmaker.Track
-
-interface RecyclerViewClickListener {
-    fun onItemClicked(track: Track)
-}
